singer aliases edit

+++ apps/server/src/api_app/controllers/update_music/update_lyric.ts
@@ -12,6 +12,13 @@ export default async ({ ctx, music, value }: Parameter) => {
     return ctx.error(ExceptionCode.PARAMETER_ERROR);
   }
 
+  const trimmedAliases: string[] = value.map((v) =>
+    v.replace(/\s+/g, ' ').trim(),
+  );
+  if (trimmedAliases.find((a) => a.length === 0)) {
+    return ctx.except(ExceptionCode.PARAMETER_ERROR);
+  }
+
   await db.run(
     `
       DELETE FROM lyric
